list  field changes in currentstock controller

+++ Libraries/HIMS.Data/DTO/Inventory/SalesReturnDetailsListDto.cs
@@ -14,7 +14,7 @@ namespace HIMS.Data.DTO.Inventory
         public string? ItemName { get; set; }
         public string? BatchNo { get; set; }
         public double? Qty { get; set; }
-        public double? MRP { get; set; }
+        public decimal? MRP { get; set; }
         public long? StoreID { get; set; }
